Handle QualifiedNameReference when selecting stream/table keys
SchemaKStream.select may get a QualifiedNameReference as the expression
in a select call. The logic that looks for and preserves the key field in
the resulting SchemaKStream needs to handle expressions of this type. It
also needs to handle cases where the key field name is fully qualified, as
happens with joins.